+++ source/dea-backend/src/config.ts
@@ -224,7 +224,7 @@ const convictSchema = {
         default: undefined,
       },
       idcenterid: {
-        doc: 'ONLY used for Identity Center, its the user id to query for users group memberships.',
+        doc: 'ONLY used for Identity Center, this is the user id to query for users group memberships.',
         format: String,
         default: undefined,
       },
@@ -255,6 +255,11 @@ const convictSchema = {
     identityStoreId: {
       doc: `identity store of your identity center instance, used for querying user's group memberships`,
       // TODO: add regex
+      format: String,
+      default: undefined,
+    },
+    identityStoreRegion: {
+      doc: `region of your identity center instance, used for querying user's group memberships`,
       format: String,
       default: undefined,
     },
@@ -385,6 +390,7 @@ export interface IdpMetadataInfo {
   readonly defaultRole: string | undefined;
   readonly groupToDeaRoleRules: GroupToDEARoleRule[];
   readonly identityStoreId: string | undefined;
+  readonly identityStoreRegion: string | undefined;
 }
 
 export interface DEAEndpointDefinition {