+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.6] - 2023-11-30
 
+### Added
+- the ability to integrate with AWS Identity Center
+- the ability to allow non Route 53 customer domains
+
 ### Fixed
+- UI Upload issue for large files
 - bump axios and crypto-js packages
 
 ## [1.0.5] - 2023-10-10