+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - when using identityStoreId, identityStoreRegion and identityStoreAccountId are also required
+- split refresh_token to a new cookie
 
 ## [1.0.6] - 2023-12-15